properly labeled create menu
