--- src/pages/Instrument.tsx
@@ -1,9 +1,0 @@
-function Instrument() {
-    return (
-        <div className="Instrument">
-            <h1>Instrument</h1>
-        </div>
-    )
-}
-
-export default Instrument
