+++ src/pages/SingleListing.tsx
@@ -1,0 +1,13 @@
+import { useParams } from 'react-router'
+
+function SingleListing() {
+    const { id } = useParams()
+    console.log(id)
+    return (
+        <div className="singleListingPage">
+            <h1>Single Listing</h1>
+        </div>
+    )
+}
+
+export default SingleListing
