+++ src/router/router.tsx
@@ -6,16 +6,18 @@ import {
 import App from '../App'
 import Homepage from '../pages/Homepage'
 import AdminLogin from '../pages/AdminLogin'
-import CreateMenu from '../pages/EditMenu'
+import CreateMenu from '../pages/CreateMenu'
 import Listings from '../pages/Listings'
+import Instrument from '../pages/SingleListing'
 
 const router = createBrowserRouter(
     createRoutesFromElements(
         <Route path="/" element={<App />}>
             <Route path="/" element={<Homepage />} />
             <Route path="/sitemaster" element={<AdminLogin />} />
-            <Route path="/editmenu" element={<CreateMenu />} />
+            <Route path="/createMenu" element={<CreateMenu />} />
             <Route path="/listings" element={<Listings />} />
+            <Route path="/listings/:id" element={<Instrument />} />
         </Route>
     )
 )
